# Slightly optimize ColumnArray::get()/operator[] (by using reserve over resize)