Changelog category (leave one):
- Not for changelog (changelog entry is not required)

Note, that simple "SELECT range(100)" will execute
ColumnArray::operator[] 14 times (most of them from
DB::checkColumnStructure())