@kitaisreal 2022-01-28
@azat actually after discussion with @alexey-milovidov it seems that main motivation is to avoid constructing Field during resize, but it seems not a big issue. Could you please somehow provide some test where this will be significant ?
If it is I will just change a lot of other similar places.