@azat 2022-01-28
>@azat actually after discussion with @alexey-milovidov it seems that main motivation is to avoid constructing Field during resize

Yes

>Could you please somehow provide some test where this will be significant ?

### before the patch

```
$ ./clickhouse benchmark --host 127.4 --query 'select range(toUInt64(1e6))'

127.4:9000, queries 129, QPS: 3.728, RPS: 3.728, MiB/s: 0.000, result RPS: 3.728, result MiB/s: 28.440.

0.000%          0.259 sec.
10.000%         0.262 sec.
20.000%         0.263 sec.
30.000%         0.264 sec.
40.000%         0.264 sec.
50.000%         0.265 sec.
60.000%         0.265 sec.
70.000%         0.266 sec.
80.000%         0.273 sec.
90.000%         0.283 sec.
95.000%         0.288 sec.
99.000%         0.300 sec.                                                                 
99.900%         0.302 sec.
99.990%         0.302 sec.
```

<details>

```
Samples: 99K of event 'cycles', 4000 Hz, Event count (approx.): 55357870155 lost: 0/0 drop: 0/3608
Overhead  Shared Object           Symbol
  29.54%  clickhouse              [.] DB::ColumnArray::operator[]
  25.72%  clickhouse              [.] DB::Field::~Field
  16.65%  clickhouse              [.] DB::Field::operator=
  13.60%  clickhouse              [.] DB::Field::operator==
   4.89%  clickhouse              [.] DB::ColumnVector<unsigned long>::operator[]          
   4.13%  clickhouse              [.] DB::Field::dispatch<DB::Field::create(DB::Field&&)::{lambda(auto:1&)#1}, DB::Field&>
```

</details>

### with this patch

```
$ ./clickhouse benchmark --host 127.3 --query 'select range(toUInt64(1e6))'

127.3:9000, queries 136, QPS: 5.028, RPS: 5.028, MiB/s: 0.000, result RPS: 5.028, result MiB/s: 38.360.     

0.000%          0.185 sec.
10.000%         0.187 sec.
20.000%         0.188 sec.
30.000%         0.190 sec.                                                                 
40.000%         0.191 sec.
50.000%         0.198 sec.
60.000%         0.201 sec.
70.000%         0.208 sec.
80.000%         0.211 sec.
90.000%         0.212 sec.
95.000%         0.214 sec.
99.000%         0.225 sec.
99.900%         0.236 sec.
99.990%         0.236 sec.
```

<details>

```
Samples: 59K of event 'cycles', 4000 Hz, Event count (approx.): 47153599213 lost: 0/0 drop: 0/587
Overhead  Shared Object       Symbol
  40.81%  clickhouse          [.] DB::Field::~Field
  19.96%  clickhouse          [.] DB::ColumnArray::operator[] # overhead reduced
  18.40%  clickhouse          [.] DB::Field::operator==
   8.65%  clickhouse          [.] DB::ColumnVector<unsigned long>::operator[]
   4.82%  clickhouse          [.] DB::Field::dispatch<DB::Field::create(DB::Field&&)::{lambda(auto:1&)#1}, DB::Field&>
```

</details>

>If it is I will just change a lot of other similar places.

It is significant only for Array/Tuple/Map, at first glance there are not a lot of them, I will update the patch.