@azat 2022-01-28
> kitaisreal merged commit e2de0df into ClickHouse:master 2 hours ago

Too early, I have not updated the patch here, anyway, here is a follow up - #34126